+++ packages/fomet_ui/lib/fomet_ui.dart
@@ -2,9 +2,11 @@
 library fomet_ui;
 
 export 'src/bottom_navigation_bar/bottom_navigation_bar.dart';
+export 'src/button/fomet_button.dart';
 export 'src/card/fomet_card.dart';
 export 'src/common/colors.dart';
 export 'src/common/constants.dart';
+export 'src/common/dimensions.dart';
 export 'src/common/typography.dart';
 export 'src/common/widgets/navigation_bar_item.dart';
 export 'src/side_navigation_bar/side_navigation_bar.dart';