+++ packages/fomet_ui/lib/src/bottom_navigation_bar/bottom_navigation_bar.dart
@@ -9,7 +9,11 @@ typedef FometBottomNavigationBarItemChanged = void Function(int);
 /// A bottom navigation bar component is generally placed at the bottom of a
 /// [Scaffold] and is used for selecting among a small number of views.
 ///
-/// Generally,
+/// The bottom navigation bar consists of multiple items in the form of text and
+/// labels laid out on top of a colored container. This widget provides quick
+/// navigation between the top-level views of the application.
+///
+/// For larger screens, prefer using [FometSideNavigationBar].
 class FometBottomNavigationBar extends StatefulWidget {
   /// The items of the navigation bar.
   final List<FometNavigationBarItem> items;
@@ -18,22 +22,33 @@ class FometBottomNavigationBar extends StatefulWidget {
   /// selected.
   final FometBottomNavigationBarItemChanged onItemTap;
 
+  /// The content above the navigation bar.
   final Widget child;
 
-  final Widget header;
+  /// An icon that is shown in the top-left corner of the page controlled by
+  /// this navigation bar widget.
+  final Widget topLeadingIcon;
 
+  /// An icon that is shown in the top-right corner of the page controlled by
+  /// this navigation bar widget.
+  ///
+  /// By default, this is set to `null`.
+  final Widget? topTrailingIcon;
+
+  /// The index of the currently visible page.
+  ///
+  /// This value changes whenever an item of [items] is tapped AND is not
+  /// already selected.
   final ValueNotifier<int> selectedIndex;
-
-  final Widget? appStatusWidget;
 
   /// Creates a [FometBottomNavigationBar] widget.
   const FometBottomNavigationBar({
     required this.items,
     required this.onItemTap,
-    required this.header,
+    required this.topLeadingIcon,
     required this.selectedIndex,
     required this.child,
-    this.appStatusWidget,
+    this.topTrailingIcon,
     super.key,
   });
 
@@ -43,15 +58,17 @@ class FometBottomNavigationBar extends StatefulWidget {
 }
 
 class _FometBottomNavigationBarState extends State<FometBottomNavigationBar> {
+  /// This is to avoid unnecessary rebuilds of the navigation bar children.
   late List<Widget> children = generateChildren();
 
+  /// Converts each [FometNavigationBarItem] into a tappable widget that will
+  /// be placed in the bottom navigation container.
   List<Widget> generateChildren() {
     return List<Widget>.generate(
       widget.items.length,
       growable: false,
       (index) {
         final element = widget.items[index];
-
         return Expanded(
           child: GestureDetector(
             onTap: () => widget.selectedIndex.value = index,
@@ -73,6 +90,8 @@ class _FometBottomNavigationBarState extends State<FometBottomNavigationBar> {
   void initState() {
     super.initState();
 
+    // This is needed to trigger the "onItemTap" callback when the user taps an
+    // item that is NOT selected.
     widget.selectedIndex.addListener(
       () => widget.onItemTap(widget.selectedIndex.value),
     );
@@ -82,6 +101,7 @@ class _FometBottomNavigationBarState extends State<FometBottomNavigationBar> {
   void didUpdateWidget(covariant FometBottomNavigationBar oldWidget) {
     super.didUpdateWidget(oldWidget);
 
+    // Rebuilds the children list only when the parent configuration changes.
     if (widget.items != oldWidget.items) {
       children = generateChildren();
     }
@@ -92,52 +112,50 @@ class _FometBottomNavigationBarState extends State<FometBottomNavigationBar> {
     return Column(
       mainAxisSize: MainAxisSize.min,
       children: [
-        // Top part
+        // Top part with the icons
         Padding(
           padding: const EdgeInsets.symmetric(
-            horizontal: 32,
-            vertical: 8,
+            horizontal: FometDimensions.space4x,
+            vertical: FometDimensions.space1x,
           ),
           child: Row(
             mainAxisAlignment: MainAxisAlignment.spaceBetween,
             children: [
-              widget.header,
-              widget.appStatusWidget ?? const SizedBox.shrink(),
+              widget.topLeadingIcon,
+              widget.topTrailingIcon ?? const SizedBox.shrink(),
             ],
           ),
         ),
         const Padding(
           padding: EdgeInsets.symmetric(
-            horizontal: 32,
+            horizontal: FometDimensions.space4x,
           ),
-          child: Divider(
-            height: 1,
-          ),
+          child: Divider(height: 1),
         ),
 
         // Body
         Expanded(
           child: Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 32),
+            padding: const EdgeInsets.symmetric(
+              horizontal: FometDimensions.space4x,
+            ),
             child: widget.child,
           ),
         ),
 
         // Separates the content from the bar
-        const Divider(
-          height: 2,
-        ),
+        const Divider(height: 2),
 
         // The actual navigation bar
         Container(
-          height: 64 + MediaQuery.of(context).padding.bottom,
+          height: FometDimensions.large + MediaQuery.of(context).padding.bottom,
           decoration: const BoxDecoration(
             color: Color(0xFFF6F6F6),
           ),
           child: Align(
             alignment: Alignment.topCenter,
             child: SizedBox(
-              height: 64,
+              height: FometDimensions.large,
               child: Row(
                 children: children,
               ),