+++ packages/fomet_ui/lib/src/common/widgets/navigation_bar_item.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:fomet_ui/fomet_ui.dart';
 
-/// An interactive button of a navigation bar component. This is used by:
-///
-///  - [FometBottomNavigationBar]
-///  - [FometSideNavigationBar]
+/// The structure of a navigation bar item. This type is used to define the
+/// children of [FometBottomNavigationBar] and [FometSideNavigationBar].
 typedef FometNavigationBarItem = ({String text, IconData iconData});