+++ packages/fomet_ui/lib/src/side_navigation_bar/side_navigation_bar.dart
@@ -6,8 +6,14 @@ import 'package:fomet_ui/src/side_navigation_bar/widget/side_navigation_item.dar
 /// user taps on a [FometNavigationBarItem].
 typedef FometSideNavigationBarItemChanged = void Function(int);
 
-/// A side navigation bar component is generally placed on the left side of a
-/// [Scaffold] and is used for selecting form a number of entries.
+/// A side navigation bar component is generally placed to the left side of the
+/// page and is used for selecting between a series of options.
+///
+/// The side navigation bar consists of multiple items in the form of text and
+/// labels enclosed by a container with animated borders. This widget provides
+/// quick navigation between the top-level views of the application.
+///
+/// For smaller screens, prefer using [FometBottomNavigationBar].
 class FometSideNavigationBar extends StatefulWidget {
   /// The items of the navigation bar.
   final List<FometNavigationBarItem> items;
@@ -16,22 +22,31 @@ class FometSideNavigationBar extends StatefulWidget {
   /// selected.
   final FometSideNavigationBarItemChanged onItemTap;
 
-  final Widget header;
+  /// An icon that is shown above all [items] widgets.
+  final Widget sidebarLeadingIcon;
 
+  /// An icon that is shown below all [items] widgets.
+  ///
+  /// By default, this is set to `null`.
+  final Widget? sidebarTrailingIcon;
+
+  /// The content next to the navigation bar.
   final Widget child;
 
+  /// The index of the currently visible page.
+  ///
+  /// This value changes whenever an item of [items] is tapped AND is not
+  /// already selected.
   final ValueNotifier<int> selectedIndex;
-
-  final Widget? appStatusWidget;
 
   /// Creates a [FometSideNavigationBar] widget.
   const FometSideNavigationBar({
     required this.items,
     required this.onItemTap,
-    required this.header,
+    required this.sidebarLeadingIcon,
     required this.child,
     required this.selectedIndex,
-    this.appStatusWidget,
+    this.sidebarTrailingIcon,
     super.key,
   });
 
@@ -40,15 +55,17 @@ class FometSideNavigationBar extends StatefulWidget {
 }
 
 class _FometSideNavigationBarState extends State<FometSideNavigationBar> {
+  /// This is to avoid unnecessary rebuilds of the navigation bar children.
   late List<Widget> children = generateChildren();
 
+  /// Converts each [FometNavigationBarItem] into a tappable widget that will
+  /// be placed in the side navigation container.
   List<Widget> generateChildren() {
     return List<Widget>.generate(
       widget.items.length,
       growable: false,
       (index) {
         final element = widget.items[index];
-
         return GestureDetector(
           onTap: () => widget.selectedIndex.value = index,
           child: SideNavigationItem(
@@ -66,6 +83,8 @@ class _FometSideNavigationBarState extends State<FometSideNavigationBar> {
   void initState() {
     super.initState();
 
+    // This is needed to trigger the "onItemTap" callback when the user taps an
+    // item that is NOT selected.
     widget.selectedIndex.addListener(
       () => widget.onItemTap(widget.selectedIndex.value),
     );
@@ -75,6 +94,7 @@ class _FometSideNavigationBarState extends State<FometSideNavigationBar> {
   void didUpdateWidget(covariant FometSideNavigationBar oldWidget) {
     super.didUpdateWidget(oldWidget);
 
+    // Rebuilds the children list only when the parent configuration changes.
     if (widget.items != oldWidget.items) {
       children = generateChildren();
     }
@@ -100,25 +120,39 @@ class _FometSideNavigationBarState extends State<FometSideNavigationBar> {
                     mainAxisSize: MainAxisSize.min,
                     children: [
                       Padding(
-                        padding: const EdgeInsets.symmetric(vertical: 36),
-                        child: widget.header,
+                        padding: const EdgeInsets.symmetric(
+                          vertical: FometDimensions.space4x,
+                        ),
+                        child: widget.sidebarLeadingIcon,
                       ),
                       ...children,
-                      if (widget.appStatusWidget != null)
+                      if (widget.sidebarTrailingIcon != null)
                         Padding(
-                          padding: const EdgeInsets.symmetric(vertical: 16),
-                          child: widget.appStatusWidget,
+                          padding: const EdgeInsets.symmetric(
+                            vertical: FometDimensions.space2x,
+                          ),
+                          child: widget.sidebarTrailingIcon,
                         ),
                     ],
                   ),
                 ),
               ),
             ),
-            const SizedBox(width: 24),
+
+            // Space
+            const SizedBox(
+              width: FometDimensions.space3x,
+            ),
+
+            // The body
             Expanded(
               child: widget.child,
             ),
-            const SizedBox(width: 8),
+
+            // Other trailing space
+            const SizedBox(
+              width: FometDimensions.space1x,
+            ),
           ],
         ),
       ),