+++ packages/fomet_ui/lib/src/side_navigation_bar/widget/side_navigation_item.dart
@@ -1,11 +1,25 @@
 import 'package:flutter/material.dart';
 import 'package:fomet_ui/fomet_ui.dart';
 
+/// A widget with an icon, a label on the right and and interactive border that
+/// is used by [FometSideNavigationBar].
+///
+/// The border color changes when [index] has the same value as [selectedIndex].
 class SideNavigationItem extends StatelessWidget {
+  /// The index of the currently selected item in the [FometSideNavigationBar]
+  /// widget.
   final ValueNotifier<int> selectedIndex;
+
+  /// This item's index.
   final int index;
+
+  /// The text next to [iconData].
   final String text;
+
+  /// The icon to the left of [text].
   final IconData iconData;
+
+  /// Creates a [SideNavigationItem] widget.
   const SideNavigationItem({
     required this.selectedIndex,
     required this.index,
@@ -18,28 +32,30 @@ class SideNavigationItem extends StatelessWidget {
   Widget build(BuildContext context) {
     return Padding(
       padding: const EdgeInsets.symmetric(
-        horizontal: 24,
-        vertical: 12,
+        horizontal: FometDimensions.space3x,
+        vertical: FometDimensions.small,
       ),
       child: ValueListenableBuilder<int>(
         valueListenable: selectedIndex,
         builder: (context, currentIndex, child) {
+          // The container around the button animates the border color when it
+          // changes.
           return AnimatedContainer(
             duration: const Duration(milliseconds: 250),
-            height: 46,
+            height: FometDimensions.space6x,
             decoration: BoxDecoration(
               border: index == currentIndex
                   ? Border.all(color: FometColors.primary)
                   : null,
-              borderRadius: BorderRadius.circular(58),
+              borderRadius: BorderRadius.circular(FometDimensions.space6x),
             ),
             child: child,
           );
         },
         child: Padding(
           padding: const EdgeInsets.symmetric(
-            horizontal: 12,
-            vertical: 6,
+            horizontal: FometDimensions.small,
+            vertical: FometDimensions.space1x,
           ),
           child: Row(
             children: [
@@ -48,19 +64,19 @@ class SideNavigationItem extends StatelessWidget {
                 builder: (context, currentIndex, _) {
                   return Icon(
                     iconData,
-                    size: 24,
+                    size: FometDimensions.space3x,
                     color: index == currentIndex
                         ? FometColors.primary
                         : Colors.grey,
                   );
                 },
               ),
-              const SizedBox(width: 8),
+              const SizedBox(width: FometDimensions.space1x),
               Expanded(
                 child: Text(
                   text,
                   style: FometTypography.light.copyWith(
-                    color: Colors.grey,
+                    color: FometColors.inactiveText,
                   ),
                 ),
               ),